refactor: ISocketHandlerの位置を変更

+++ pkg/infra/socket/ISocketHandler.hpp
@@ -3,6 +3,7 @@
 
 #include <poll.h>
 #include <string>
+#include <sys/types.h>
 
 class ISocketHandler {
 public:
@@ -14,7 +15,7 @@ public:
 
   virtual int acceptConnection(struct sockaddr_in *clientAddr) = 0;
   virtual void closeConnection(int &targetSocket) = 0;
-  virtual int sendMsg(const std::string &message, int &targetSocket) = 0;
+  virtual ssize_t sendMsg(const std::string &message, int &targetSocket) = 0;
   virtual std::string receiveMsg(const int &targetSocket) = 0;
 
   virtual const int &getServerSocket() const = 0;
